ocf_kubernetes: Allow ocfroot to use kubectl
Give ocfroot read access to /etc/kubernetes/admin.conf, which has the
client cert required to authenticate as a "kubernetes-admin". Now
ocfroot can administer the cluster without becoming the root user.

Create a smaller helper script called 'ocf-kubectl' that sets the
KUBECONFIG environment variable and proxies the rest of the arguments to
kubectl.