kubernetes: stop nginx from forwarding all vhosts
Previously, the front-end nginx would forward any request (with any
host header) to the Kubernetes ingress controller. This commit gets rid
of that catch-all server block, and only forwards vhosts with the
http_vhost declaration (previously http_redirect). This is more in-line
with how we handled Marathon vhosts.

This can be used to create "hidden" Kubernetes vhosts that aren't
accessible publicly, which we need for snmp_exporter.