New CUPS settings, to have PDFs rasterized before printing
This changes CUPS settings, so print jobs first get processed locally, and then
get passed onto printhost. A Tea4CUPS filter rasterizes all PDF files.

This allows us to remove the PDF rasterization that occurs when opening a PDF
on the desktops, which is inconvenient for those who just want to read a PDF
and not actually print it out.