Further removing old PDF rasterization setup
To ensure that the old setup is gone, I am removing the files
associated with the old raster-upon-open setup. IMPORTANT: this is
meant to be TEMPORARY only; if my pull request gets merged, I will
create a new pull request to remove this code once it has served its
purpose.