Tweaked raster-filter if statements
I've changed all if statements to use the if [ condition ] style, and
removed the check that "$error" is not empty before sending an email,
since having convert fail implies an error message, which should be
emailed out. (Even with no error message, an email can still be sent,
indicating an error with the conversion.)

Note that I've also removed the -e option when setting Bash script
options, which allows me to better check the error status of convert.