Don't send debconf emails on desktops
These emails aren't really useful and happen whenever a new desktop is
provisioned or video drivers are updated. For instance, this has emailed
in the past for the libcuda1, xserver-xorg-video-nvidia, and
libgl1-nvidia-glx packages and just warns about a mismatch in video
drivers and a need to reboot the desktop, which we do after provisioning
them anyway.