Remove a couple duplicated files left over from jessie
In 28f3e7d1ebe675cb31b731ed6f2cfb38a5b813cf I thought these just had
"stretch" added to their name because there was a previous jessie
version, but it turns out that version was still in the repo and was in
the "more correct" location for the config file. Since we don't have any
jessie hosts any more, let's update the more authoritative source with
the correct contents for stretch and remove the (now unnecessary)
stretch versions of these files.