Move desktop files out of Xsession to /etc/skel (fixes #389)
This should be a decent amount cleaner and easier to maintain, and we
don't need the full conditional power that comes with using a script to
generate the desktop files since they are all pretty simple. I
simplified the remote file launcher since we assume remote files
appeared properly anyway, and if they didn't, then something else is
wrong and we shouldn't just hide the desktop icon.