ocf: store ED25519 hostkeys in Puppet
Right now, ED25519 keys are generated when sshd is installed, but
they are not stored in Puppet, so when the host is rebuilt, the
ED25519 keys might change, causing host verification failed warnings
for clients which chose to use that key.
Store the keys in Puppet to prevent these kinds of problems.