ocf_irc: Add SASL option for IRC connections
This adds the SASL module to anope and inspircd, and tells inspircd to
use anope for SASL. Note that SASL authentication also requires the
capability negotiation protocol for clients.

Although SASL is part of the ircv3 spec, most clients and large IRC
networks have SASL support.

We deliberately do not support the dh_aes (or its predecessor dh_blowfish)
mechanisms. These mechanisms are for users that want to send their
username and passwords encrypted to the server over an insecure
connection. Since the OCF only supports connecting over TLS there is no
point.