ocf_backups: split backup before encrypting
This makes it easier to recover from an encrypted backup. I don't have
a machine with 5T, so I would have to put my private keys on an OCF
server to recover from backups (since the whole ~5T image is
encrypted). This change encrypts each 5G part separately, which allows
me to download a chunk, decrypt it, move it to a server, then do the
next chunk, without having them all on my machine at the same time.