fix systemd not having permissions to read procfs (#1218)
It seems `gid=` can only be specified once for procfs mount options, so
the `capptrace` group was being ignored entirely; switch to using
`ocfstaff` as the designated group, which is ugly but should work. In
addition, systemd now also needs to read procfs for running the user
daemon, so add a drop-in for `user@.service` as well.