feat(desktop): kill locked-and-idle sessions
Add a daemon to kill all locked sessions after they have been
locked for a certain amount of time (currently ~3 minutes). Currently
this is implemented by polling every minute, so if a user unlocks and
locks between ticks, it is ignored. Also, the state is never persisted,
so when the daemon restarts (e.g. puppet updates it), the timers get
reset.

This also applies to staff/opstaff, but at a much longer delay (4
hours).