Added code to clear google chrome cache and config upon xsession cleanup (ie user logging out).
This should fix an issue with chrome cache and
config taking up too much space to the point
of crashing the desktops.

Does remove persistance of config between user
sessions, but it is chrome specific, and we
do not explicitly support such persistance
anyways.

Tested on my laptop with lightdm and works.
Fairly simple and low risk. Only bash code.
2 lines.
``